misra shadow: resolve 10.1 rule violations

+++ libraries/aws/shadow/src/aws_iot_shadow_api.c
@@ -349,7 +349,7 @@ static AwsIotShadowError_t _setCallbackCommon( IotMqttConnection_t mqttConnectio
     }
 
     /* Check parameters. */
-    status = _validateThingNameFlags( ( _shadowOperationType_t ) ( type + SHADOW_OPERATION_COUNT ),
+    status = _validateThingNameFlags( ( _shadowOperationType_t ) ( ( int32_t ) type + SHADOW_OPERATION_COUNT ),
                                       pThingName,
                                       thingNameLength,
                                       0,
@@ -628,7 +628,7 @@ static void _callbackWrapperCommon( _shadowCallbackType_t type,
     IotMutex_Unlock( &_AwsIotShadowSubscriptionsMutex );
 
     /* Set the callback type. Shadow callbacks are enumerated after the operations. */
-    callbackParam.callbackType = ( AwsIotShadowCallbackType_t ) ( type + SHADOW_OPERATION_COUNT );
+    callbackParam.callbackType = ( AwsIotShadowCallbackType_t ) ( ( int32_t ) type + SHADOW_OPERATION_COUNT );
 
     /* Set the remaining members of the callback param. */
     callbackParam.mqttConnection = pMessage->mqttConnection;
